+++ internal/provider/provider.go
@@ -178,5 +178,7 @@ func (p *clearpassProvider) DataSources(ctx context.Context) []func() datasource
 		NewEnforcementProfilesDataSource,
 		NewLocalUserDataSource,
 		NewLocalUsersDataSource,
+		NewRoleDataSource,
+		NewRolesDataSource,
 	}
 }